fix: Move widget to its own module and partially fix for new labels
